refactor: consolidate game utilities into boardUtils
- Move utility functions from constants.js to boardUtils.js
- Remove unused REPLAY_SIDES enum (code uses string literals directly)
- Update imports in gameManager.js
- Update CLAUDE.md documentation
