+++ README.md
@@ -1,6 +1,5 @@
 # Kifu-oku (棋譜憶)
 
-
 Kifu-oku (棋譜憶) is a coined Japanese term combining 棋譜 (kifu), meaning "game record" with 憶 (oku), meaning "to remember" or "recollect", the same character found in 記憶 (kioku, "memory"). Together, 棋譜憶 suggests "recalling game records" or "kifu memory."
 
 While Kifu-oku (棋譜憶) is not a word you'll find in a dictionary, the construction follows natural Japanese compounding patterns and its meaning is immediately apparent to readers familiar with the characters.
