+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kifuoku",
-  "version": "0.3.1",
+  "version": "0.4.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
