+++ src/game/constants.js
@@ -20,47 +20,7 @@ export const PHASES = {
   COMPLETE: 'complete'
 }
 
-export const REPLAY_SIDES = {
-  BOTH: null,
-  BLACK: 'B',
-  WHITE: 'W'
-}
-
 export const HINT_TYPES = {
   QUADRANT: 'quadrant',
   EXACT: 'exact'
-}
-
-export function getQuadrant(x, y, boardSize = DEFAULT_BOARD_SIZE) {
-  const midpoint = boardSize / 2
-  const horizontal = x < midpoint ? 'left' : 'right'
-  const vertical = y < midpoint ? 'upper' : 'lower'
-  return `${vertical} ${horizontal}`
-}
-
-export function getQuadrantVertices(quadrant, boardSize = DEFAULT_BOARD_SIZE) {
-  const vertices = []
-  const midpoint = Math.floor(boardSize / 2)
-
-  const [vertical, horizontal] = quadrant.split(' ')
-  const xRange = horizontal === 'left' ? [0, midpoint - 1] : [midpoint, boardSize - 1]
-  const yRange = vertical === 'upper' ? [0, midpoint - 1] : [midpoint, boardSize - 1]
-
-  for (let y = yRange[0]; y <= yRange[1]; y++) {
-    for (let x = xRange[0]; x <= xRange[1]; x++) {
-      vertices.push([x, y])
-    }
-  }
-
-  return vertices
-}
-
-export function colorToSign(color) {
-  if (color === 'B') return 1
-  if (color === 'W') return -1
-  return 0
-}
-
-export function randomInt(min, max) {
-  return Math.floor(Math.random() * (max - min + 1)) + min
 }
